api/requirements.txt
alpaca-trade-api===0.51.0
asgiref===3.2.10
certifi===2020.6.20
chardet===3.0.4
Django===3.1.2
django-debug-toolbar===3.1.1
django-environ===0.4.5
django-filter===2.4.0
djangorestframework===3.12.1
factory-boy===3.1.0
Faker===4.14.0
idna===2.10
Markdown===3.3.3
numpy===1.19.4
pandas===1.1.4
pip===20.2.2
psycopg2===2.8.6
python-dateutil===2.8.1
pytz===2020.1
requests===2.24.0
setuptools===49.6.0
six===1.15.0
sqlparse===0.4.1
text-unidecode===1.3
urllib3===1.25.11
websocket-client===0.57.0
websockets===8.1
wheel===0.35.1